+++ packages/htlc/src/schemas/index.ts
@@ -1,4 +1,4 @@
-export * from './UTXOAdminRefundPubKeyRedeemScript';
-export * from './UTXOAdminRefundPubKeyHashRedeemScript';
-export * from './UTXOPubKeyRedeemScript';
-export * from './UTXOPubKeyHashRedeemScript';
+export * from './utxo-admin-refund-pubkey-redeemscript';
+export * from './utxo-admin-refund-pubkeyhash-redeemscript';
+export * from './utxo-pubkeyhash-redeemscript';
+export * from './utxo-pubkey-redeemscript';
