+++ packages/htlc/src/schemas/utxo-admin-refund-pubkeyhash-redeemscript.ts
@@ -1,7 +1,7 @@
 import { DecompiledOpCode, SwapError } from '@radar/redshift-types';
 
-export const uTXOPubKeyHashRedeemScriptSchema = {
-  id: '/UTXOPubKeyHashRedeemScriptSchema',
+export const uTXOAdminRefundPubKeyHashRedeemScriptSchema = {
+  id: '/UTXOAdminRefundPubKeyHashRedeemScriptSchema',
   type: 'array',
   items: [
     {
@@ -45,6 +45,30 @@ export const uTXOPubKeyHashRedeemScriptSchema = {
       errorMessage: SwapError.EXPECTED_OP_ELSE,
     },
     {
+      type: 'string',
+      enum: [DecompiledOpCode.OP_DUP],
+      errorMessage: SwapError.EXPECTED_OP_DUP,
+    },
+    {
+      type: 'string',
+      enum: [DecompiledOpCode.OP_HASH160],
+      errorMessage: SwapError.EXPECTED_OP_HASH160,
+    },
+    {
+      // decompiledRefundHashRipemd160
+      type: 'string',
+    },
+    {
+      type: 'string',
+      enum: [DecompiledOpCode.OP_EQUAL],
+      errorMessage: SwapError.EXPECTED_OP_EQUAL,
+    },
+    {
+      type: 'string',
+      enum: [DecompiledOpCode.OP_NOTIF],
+      errorMessage: SwapError.EXPECTED_OP_NOTIF,
+    },
+    {
       // decompiledTimeLockValue
       type: 'string',
     },
@@ -55,6 +79,11 @@ export const uTXOPubKeyHashRedeemScriptSchema = {
         DecompiledOpCode.OP_CHECKLOCKTIMEVERIFY,
       ],
       errorMessage: SwapError.INVALID_TIMELOCK_METHOD,
+    },
+    {
+      type: 'string',
+      enum: [DecompiledOpCode.OP_ENDIF],
+      errorMessage: SwapError.EXPECTED_OP_ENDIF,
     },
     {
       type: 'string',
@@ -74,7 +103,8 @@ export const uTXOPubKeyHashRedeemScriptSchema = {
     {
       // decompiledRefundPublicKeyHash
       type: 'string',
-      length: 40,
+      length: 66,
+      errorMessage: 'Invalid Refund Public Key',
     },
     {
       type: 'string',
@@ -92,4 +122,4 @@ export const uTXOPubKeyHashRedeemScriptSchema = {
       errorMessage: SwapError.EXPECTED_OP_CHECKSIG,
     },
   ],
-} as any;
+};
